+++ server/commands/cycle.js
@@ -1,4 +1,6 @@
-/* global graphQLFetcher, notifyUser, commandsConfig */
+/* global graphQLFetcher, notifyUser, usageFormat, usageMessage, commandFuncs */
+
+const {cycle} = Npm.require('@learnersguild/game-cli')
 
 function invokeUpdateCycleStateAPI(state, lgJWT) {
   const baseURL = process.env.NODE_ENV === 'development' ? 'http://game.learnersguild.dev' : 'https://game.learnersguild.org'
@@ -13,7 +15,6 @@ function invokeUpdateCycleStateAPI(state, lgJWT) {
 function handleUpdateCycleStateCommand(commandInfo, state, msg) {
   try {
     const {lgJWT, lgUser} = Meteor.user().services.lgSSO
-    console.log({lgUser})
     if (!lgJWT || lgUser.roles.indexOf('moderator') < 0) {
       throw new Error('You are not a moderator.')
     }
@@ -29,37 +30,29 @@ function handleUpdateCycleStateCommand(commandInfo, state, msg) {
   }
 }
 
-function showUsage(rid) {
-  notifyUser(rid, `**USAGE:**
-  \`/cycle launch\` - Form teams and launch the cycle
-  \`/cycle retro\` - Start the retrospective
-  \`/cycle help\` - Show this handy usage information
-`)
-}
-
-commandsConfig.cycle.onInvoke = (command, commandParamStr, commandInfo) => {
-  const subcommands = commandParamStr.split(/\s+/).filter(subcommand => subcommand.length > 0)
-  if (subcommands.length > 0) {
-    const subcommand = subcommands[0]
-    switch (subcommand) {
-      case 'launch': {
-        handleUpdateCycleStateCommand(commandInfo, 'PRACTICE', '🚀  Initiating Launch... stand by.')
-        break
-      }
-      case 'retro': {
-        handleUpdateCycleStateCommand(commandInfo, 'RETROSPECTIVE', '🤔  Initiating Retrospective... stand by.')
-        break
-      }
-      case '--help':
-      case 'help': {
-        showUsage(commandInfo.rid)
-        break
-      }
-      default: {
-        notifyUser(commandInfo.rid, '**ERROR:** Invalid action for `/cycle`. Try `/cycle --help` for usage.')
-      }
+commandFuncs.cycle = (command, commandParamStr, commandInfo) => {
+  let args
+  try {
+    args = cycle.parse(commandParamStr.split(/\s+/))
+  } catch (error) {
+    return notifyUser(commandInfo.rid, usageMessage(error))
+  }
+  const usageText = cycle.usage(args)
+  if (usageText) {
+    notifyUser(commandInfo.rid, usageFormat(usageText))
+  } else if (args._.length === 1) {
+    const subcommandFuncs = {
+      launch: () => handleUpdateCycleStateCommand(commandInfo, 'PRACTICE', '🚀  Initiating Launch... stand by.'),
+      retro: () => handleUpdateCycleStateCommand(commandInfo, 'RETROSPECTIVE', '🤔  Initiating Retrospective... stand by.'),
+      status: () => notifyUser(commandInfo.rid, 'Sorry -- `status` is not yet implemented.'),
+    }
+    const subcommandFunc = subcommandFuncs(args._[0])
+    if (subcommandFunc) {
+      subcommandFunc()
+    } else {
+      notifyUser(commandInfo.rid, '**ERROR:** Invalid action for `/cycle`. Try `/cycle --help` for usage.')
     }
   } else {
-    showUsage(commandInfo.rid)
+    notifyUser(commandInfo.rid, usageFormat(cycle.usage()))
   }
 }