+++ HISTORY.md
@@ -1,8 +1,11 @@
 # History
 
+#### 0.6.0
+Integrated `game-cli` npm package for better CLI parsing.
+
 #### 0.5.0
-- Adding `/cycle retro` command
-- Adding `/cycle help` command
+- Added `/cycle retro` command
+- Added `/cycle help` command
 
 #### 0.4.4
 - better error message when API returns error
@@ -13,7 +16,7 @@
 GraphQL mutation must select result to return
 
 #### 0.4.0
-Introduced the /cycle launch command
+Introduced the `/cycle launch` command
 
 #### 0.3.5
 Moving validation for votes to server